Absolute distance heuristic calculation

+++ src/environment/HeuristicMoveInterface.java
@@ -3,5 +3,5 @@ package environment;
 public interface HeuristicMoveInterface<M extends HeuristicMoveInterface<M, I>, I>
         extends MoveInterface<M, I> {
 
-    Integer getTargetHeuristic();
+    Double getTargetHeuristic();
 }
